:sparkles: Feat: 타이머 api 연결 및 컴포넌트 분리
